requirements.txt
fastapi==0.115.5
httpx==0.28.1
fastapi-mail==1.4.1
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
python-multipart==0.0.12
email-validator==2.2.0
